+++ src/tomate/db_types/data_disk.py
@@ -316,7 +316,6 @@ class DataDisk(DataBase):
         """
         if not self.filegroups:
             raise IndexError("No filegroups in database.")
-        self.check_regex()
         self.check_scanning_functions()
         for fg in self.filegroups:
             fg.scan_files()
@@ -450,15 +449,6 @@ class DataDisk(DataBase):
             if all(s > 0 for s in intersect):
                 raise ValueError("Duplicate values in filegroups {} and {}"
                                  .format(fg1.name, fg2.name))
-
-    def check_regex(self):
-        """Check if a pregex has been added where needed.
-
-        :raises AttributeError: If regex is empty and there is no file_override.
-        """
-        for fg in self.filegroups:
-            if fg.regex == '' and fg.file_override == '':
-                raise AttributeError(f"Filegroup '{fg.name}' is missing a regex.")
 
     def check_scanning_functions(self):
         """Check if CoordScan have scanning functions set.