+++ changelog.md
@@ -1,5 +1,8 @@
+- [2020-09-27] Add process to scan coordinates objects.
+- [2020-09-27] Replace `filegroup.file_override` by `constructor.set_file_override`
 - [2020-09-27] Send variables from a post loading function selection being loaded as
   positionial arguments.
+
 # v2.1.2
 
 - [2020-09-27] Remove shape check when supplying data to create plot object.